trunner: add erasing flash before flashing for zynq7000-zedboard
Change, which was needed with this commit was sending only \r after passed command (plo parser needs only this).
It was necessary, because erase command checks whether first available character is Y (coz of YES!).
If we were sending erase\r\n, \n would be this character and erase would fail.

JIRA: PD-268