exclude waitpid test on zynqmp-qemu target
Due to #1450 issue the --thread=single qemu argument doesn't allow this test to be run.
The infinite for loop in one thread makes this target extremely slow and a few minutes of timeout would be needed.
Knowing that the hw zynqmp target will be added, it seems to be the best solution for now.
Removing the --thread=single option is not possible, because it would make this target much slower in the other scenarios.

JIRA: CI-594